fix(game): do not update the game mod_log with problem data
Note:

Even though a `Problem` can be inside the `Game`, we do not consider the `Game` to have been modified if a `Problem` was modified. A target is considered to have been modified only if a non-trivial `modify()` or `_log()` hook is called when the mod attempts to modify the target.